+++ packages/frontend/src/features/sync/google-sheets/components/sync-modal/sync-modal-delete.tsx
@@ -1,9 +1,9 @@
 import { NonIdealState, Spinner } from '@blueprintjs/core';
 import { Button, Group, Stack, Text } from '@mantine/core';
 import { useCallback, useEffect } from 'react';
-import ErrorState from '../../../../../../components/common/ErrorState';
-import { useScheduler } from '../../../../../../hooks/scheduler/useScheduler';
-import { useSchedulersDeleteMutation } from '../../../../../../hooks/scheduler/useSchedulersDeleteMutation';
+import ErrorState from '../../../../../components/common/ErrorState';
+import { useScheduler } from '../../../../../hooks/scheduler/useScheduler';
+import { useSchedulersDeleteMutation } from '../../../../../hooks/scheduler/useSchedulersDeleteMutation';
 import {
     SyncWithGoogleSheetsModalAction,
     useSyncWithGoogleSheetsModal,
