+++ packages/frontend/src/features/sync/google-sheets/components/sync-modal/sync-modal-form.tsx
@@ -16,13 +16,13 @@ import {
 import { IconCirclesRelation } from '@tabler/icons-react';
 import { FC, useEffect } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
-import ErrorState from '../../../../../../components/common/ErrorState';
-import MantineIcon from '../../../../../../components/common/MantineIcon';
-import CronInput from '../../../../../../components/ReactHookForm/CronInput';
-import { useChartSchedulerCreateMutation } from '../../../../../../hooks/scheduler/useChartSchedulers';
-import { useScheduler } from '../../../../../../hooks/scheduler/useScheduler';
-import { useSchedulersUpdateMutation } from '../../../../../../hooks/scheduler/useSchedulersUpdateMutation';
-import { isInvalidCronExpression } from '../../../../../../utils/fieldValidators';
+import ErrorState from '../../../../../components/common/ErrorState';
+import MantineIcon from '../../../../../components/common/MantineIcon';
+import CronInput from '../../../../../components/ReactHookForm/CronInput';
+import { useChartSchedulerCreateMutation } from '../../../../../hooks/scheduler/useChartSchedulers';
+import { useScheduler } from '../../../../../hooks/scheduler/useScheduler';
+import { useSchedulersUpdateMutation } from '../../../../../hooks/scheduler/useSchedulersUpdateMutation';
+import { isInvalidCronExpression } from '../../../../../utils/fieldValidators';
 import {
     SyncWithGoogleSheetsModalAction,
     useSyncWithGoogleSheetsModal,
